move searchcomponent
